requirements.txt
new-package==0.0.1
selenium==4.15.2
beautifulsoup4==4.12.2
webdriver-manager==4.0.1
google-generativeai==0.3.1
schedule==1.2.0
python-dotenv==1.0.0
pandas>=2.2.0
fpdf2==2.7.6
Pillow==10.1.0
